Consolidate record extraction script to re_id_tool.py
